Create flake
Adds a flake to `/nix`. Contains outputs for the package itself as well
as a devShell, overlay, and the existing NixOS module. This required
modifying the NixOS module to use `pkgs.kmonad` as the default package
instead of `import`ing `/nix/default.nix`, as that derivation is
impure due to its pinning of Nixpkgs.

Also removes a deprecated use of `stdenv.lib` in `/nix/kmonad.nix` and
fixes a typo in /nix/nixos-module.